Done. Now comes the cleanup!
