+++ src/dcegm/final_period.py
@@ -78,64 +78,3 @@ def final_period_wrapper(
         final_value,
         marginal_utilities_choices,
     )
-
-
-def aggregate_marg_utils_exp_values(
-    final_value_state_choice: jnp.ndarray,
-    state_times_state_choice_mat: jnp.ndarray,
-    marg_util_state_choice: jnp.ndarray,
-    sum_state_choices_to_state: jnp.ndarray,
-    taste_shock_scale: float,
-    income_shock_weights: jnp.ndarray,
-) -> Tuple[jnp.ndarray, jnp.ndarray]:
-    """Computes the aggregate marginal utilities and expected values.
-
-    Args:
-        final_value_state_choice (jnp.ndarray): 2d array of shape
-            (n_states * n_choices, n_income_shocks) of the value function for all
-            states, choices, and income shocks.
-        marg_util_state_choice (jnp.ndarray): 2d array of shape
-            (n_states * n_choices, n_income_shocks) of the marginal utility of
-            consumption for all states, choices, and income shocks.
-        sum_state_choices_to_state (jnp.ndarray): 2d array of shape
-            (n_states, n_states * n_choices) with state_space size  with ones, where
-            state-choice belongs to state.
-        taste_shock_scale (float): The taste shock scale.
-        income_shock_weights (jnp.ndarray): 1d array of shape (n_stochastic_quad_points)
-
-    Returns:
-        tuple:
-
-        - marginal_utils_draws (np.ndarray): 1d array of shape
-        (n_states, n_savings, n_income_shocks,) of the aggregate marginal utilities.
-        - max_exp_values_draws (np.ndarray): 1d array of shape
-        (n_states, n_savings, n_income_shocks,) of the aggregate expected values.
-
-    """
-    max_value_per_state = jnp.take(
-        final_value_state_choice,
-        state_times_state_choice_mat,
-        axis=0,
-    ).max(axis=1)
-    rescale_value = jnp.tensordot(
-        sum_state_choices_to_state, max_value_per_state, axes=(0, 0)
-    )
-
-    exp_value = jnp.exp((final_value_state_choice - rescale_value)) ** (
-        1 / taste_shock_scale
-    )
-    sum_exp = jnp.tensordot(sum_state_choices_to_state, exp_value, axes=(1, 0))
-
-    max_exp_values_draws = max_value_per_state + taste_shock_scale * jnp.log(sum_exp)
-    marginal_utils_draws = jnp.divide(
-        jnp.tensordot(
-            sum_state_choices_to_state,
-            jnp.multiply(exp_value, marg_util_state_choice),
-            axes=(1, 0),
-        ),
-        sum_exp,
-    )
-    return (
-        marginal_utils_draws @ income_shock_weights,
-        max_exp_values_draws @ income_shock_weights,
-    )
